Refactor:JS functions to AlpineJS format
